Fix tournament clearing while making a grid
The game will try to clear the active tournament grid if it's from an
old tournament. With T2 this can be annoying as it often only gets
active when building a grid (and then getting that cleared after 0-30
seconds).
Change it to clear both the active and non-active tournament, so grids
should be cleared within 30 seconds of tournament running.
Also update tid when manually clearing.